v0.2 Release - Code Refactored Completely
- The configuration file has been split into different parts for the
email which now has "email_addresses" and "email_settings to differentiate
between server settings and address settings. It also includes a subject
field for the emails.

- MIT licesne included.

- Unit tests added for all functions that are not completely dependent on
external libraries.

- Emails are now sent in a goroutine. They were the slowest part of the
whole monitoring process, so it should now interfere less with the monitoring

- Alerts are now sensitive to being active or not. Once an alert is sent
it is considered active, and then another "recovery" email is sent once
the machine has dropped below the threshold.